move pojo to correct package
