+++ customer/src/main/java/org/oncors/model/Address.java
@@ -1,4 +1,4 @@
-package org.oncors.dao;
+package org.oncors.model;
 
 import lombok.Getter;
 import lombok.Setter;
